sodimac/black_box/casos_de_uso.py
# Caso de prueba

# Caso de prueba 1: Agregar Producto al Inventario


def agregar_producto_al_inventario(nuevo_producto):
    """
    Simula la adición de un nuevo producto al inventario con valores válidos.
    Verifica que el producto se haya agregado correctamente al inventario.
    """
    # Implementación simulada para caja negra
    producto_agregado = True  # Simulamos que el producto se agrega con éxito.
    return producto_agregado

# Caso de prueba 2: Crear Pedido de Producto


def crear_pedido_producto(datos_validos):
    """
    Simula la creación de un pedido de un producto.
    Verifica que el pedido se haya creado correctamente y afecte el inventario.
    """
    # Implementación simulada para caja negra
    pedido_creado = datos_validos
    inventario_afectado = datos_validos
    return datos_validos

# Caso de prueba 3: Notificar Niveles Mínimos de Inventario


def notificar_niveles_minimos_inventario(inventario_alcanza_minimos):
    """
    Simula la notificación a los administradores cuando el inventario alcanza niveles mínimos.
    Verifica que se envíe una notificación a los administradores.
    """
    # Implementación simulada para caja negra
    notificacion_enviada = True
    return notificacion_enviada

# Caso de prueba 4: Ejecutar Algoritmo de Reabastecimiento


def ejecutar_algoritmo_reabastecimiento():
    """
    Simula la ejecución del algoritmo de reabastecimiento manualmente.
    Verifica que el algoritmo determine de manera adecuada los productos que necesitan ser reabastecidos.
    """
    # Implementación simulada para caja negra
    productos_reabastecidos = True
    return productos_reabastecidos

# Caso de prueba 5: Validar Cantidad de Producto en Orden de Compra


def validar_cantidad_producto_orden_compra(cantidad):
    """
    Simula la validación de la cantidad de productos en una orden de compra.
    Verifica que el sistema muestre un mensaje de error cuando se ingresa una cantidad no válida.
    """
    # Implementación simulada para caja negra
    cantidad_valida = cantidad > 0
    mensaje_error = "Cantidad no válida" if not cantidad_valida else ""
    return cantidad_valida, mensaje_error

# Caso de prueba 6: Calcular Disponibilidad de Productos


def calcular_disponibilidad_productos(stock_disponible, pedidos_pendientes):
    """
    Simula el cálculo de la disponibilidad de productos en función del stock y los pedidos pendientes.
    Verifica que la disponibilidad se calcule correctamente.
    """
    # Implementación simulada para caja negra
    disponibilidad = stock_disponible - pedidos_pendientes
    return disponibilidad

# Caso de prueba 7: Aplicar Descuento por Nivel de Stock


def aplicar_descuento_por_nivel_stock(nivel_stock):
    """
    Simula la aplicación de descuentos a productos en función de su nivel de stock.
    Verifica que el descuento se aplique correctamente al precio del producto.
    """
    # Implementación simulada para caja negra
    descuento_aplicado = nivel_stock >= 10
    return descuento_aplicado

# Caso de prueba 8: Sincronización con Proveedor


def sincronizar_con_proveedor():
    """
    Simula la sincronización con el proveedor para mantener actualizado el inventario.
    Verifica que la sincronización se complete con éxito y que los datos del inventario se actualicen.
    """
    # Implementación simulada para caja negra
    sincronizacion_exitosa = True
    datos_actualizados = True
    return sincronizacion_exitosa, datos_actualizados

# Caso de prueba 9: Confirmación de Pedido


def confirmar_pedido():
    """
    Simula la confirmación de un pedido de productos.
    Verifica que el pedido se confirme correctamente y se actualice el estado del inventario.
    """
    # Implementación simulada para caja negra
    pedido_confirmado = True
    estado_inventario_actualizado = True
    return pedido_confirmado, estado_inventario_actualizado

# Caso de prueba 10: Acceso de Usuario no Autorizado


def acceso_usuario_no_autorizado():
    """
    Simula el intento de acceso a funciones de modificación del inventario con un usuario no autorizado.
    Verifica que se deniegue el acceso y se muestre un mensaje de error.
    """
    # Implementación simulada para caja negra
    acceso_denegado = True
    mensaje_error = "Acceso no autorizado" if not acceso_denegado else ""
    return acceso_denegado, mensaje_error

# Caso de prueba 11: Manejo de Error en Comunicación con el Proveedor


def manejar_error_comunicacion_proveedor(error_simulado):
    """
    Simula el manejo de errores de comunicación con el proveedor.
    Verifica que el sistema registre el error y muestre un mensaje de error al usuario.
    """
    # Implementación simulada para caja negra
    error_registrado = True
    mensaje_error = "Error de comunicación con el proveedor" if error_simulado else ""
    return error_registrado, mensaje_error

# Caso de prueba 12: Publicación de Reseña de Producto


def publicar_resena_producto(resena):
    """
    Simula la publicación de una reseña de producto por parte de los usuarios.
    Verifica que la reseña se publique correctamente y esté disponible en la página del producto.
    """
    # Implementación simulada para caja negra
    reseña_publicada = True
    return reseña_publicada

# Caso de prueba 13: Prueba de Compatibilidad con Navegadores


def probar_compatibilidad_navegadores():
    """
    Simula la prueba de compatibilidad con diferentes navegadores web.
    Verifica que las funciones del sistema de inventario se comporten correctamente en cada navegador.
    """
    # Implementación simulada para caja negra
    compatibilidad_correcta = True
    return compatibilidad_correcta

# Caso de prueba 14: Búsqueda de producto por Nombre


def buscar_producto_por_nombre(nombre_buscado, resultados_esperados):
    """
    Simula la funcionalidad de búsqueda de productos por nombre en el sistema de inventario.
    Verifica que el sistema muestre el producto buscado en los resultados de búsqueda.
    """
    # Implementación simulada para caja negra
    producto_encontrado = nombre_buscado in resultados_esperados
    return producto_encontrado

# Caso de prueba 15: Compra de producto con descuento


def comprar_producto_con_descuento(nivel_stock, precio_original):
    """
    Simula la compra de un producto con descuento en el proceso de compra.
    Verifica que el precio del producto refleje el descuento en el resumen de compra.
    """
    # Implementación simulada para caja negra
    precio_final = precio_original - 10 if nivel_stock >= 10 else precio_original
    return precio_final

# Caso de prueba 16: Prueba de notificación al alcanzar niveles mínimos de inventario


def probar_notificacion_niveles_minimos_inventario(inventario_alcanza_minimos):
    """
    Simula la notificación al alcanzar niveles mínimos de inventario.
    Verifica que el sistema envíe una notificación a los administradores.
    """
    # Implementación simulada para caja negra
    notificacion_enviada = inventario_alcanza_minimos
    return notificacion_enviada
